New apiserver watcher package
Rather than having all the watchers in one file, we move the watchers
into their own dedicated watchers package. This makes is much easier
to reason about and navigate.

The watchers all attempt to first locate the watcher from the watcher
registry, upon failing that with a not found, will fallback to the
resources. As resources are going to be deprecated we'll hopefully
seeing the usage of that being removed over time.
