+++ apiserver/watchers/allwatcher_test.go
@@ -1,7 +1,7 @@
 // Copyright 2020 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package apiserver
+package watchers
 
 import (
 	"github.com/golang/mock/gomock"
