+++ apiserver/watchers/stringswatcher.go
@@ -1,0 +1,70 @@
+// Copyright 2023 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package watchers
+
+import (
+	"github.com/juju/errors"
+
+	apiservererrors "github.com/juju/juju/apiserver/errors"
+	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/rpc/params"
+	"github.com/juju/juju/state"
+)
+
+// srvStringsWatcher defines the API for methods on a state.StringsWatcher.
+// Each client has its own current set of watchers, stored in resources.
+// srvStringsWatcher notifies about changes for all entities of a given kind,
+// sending the changes as a list of strings.
+type srvStringsWatcher struct {
+	watcherCommon
+	watcher state.StringsWatcher
+}
+
+func NewStringsWatcher(context facade.Context) (facade.Facade, error) {
+	var (
+		id              = context.ID()
+		auth            = context.Auth()
+		watcherRegistry = context.WatcherRegistry()
+		resources       = context.Resources()
+	)
+
+	// TODO(wallyworld) - enhance this watcher to support
+	// anonymous api calls with macaroons.
+	if auth.GetAuthTag() != nil && !isAgentOrUser(auth) {
+		return nil, apiservererrors.ErrPerm
+	}
+
+	watcher, err := GetWatcherByID(watcherRegistry, resources, id)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	stringsWatcher, ok := watcher.(state.StringsWatcher)
+	if !ok {
+		return nil, apiservererrors.ErrUnknownWatcher
+	}
+	return &srvStringsWatcher{
+		watcherCommon: newWatcherCommon(context),
+		watcher:       stringsWatcher,
+	}, nil
+}
+
+// Next returns when a change has occurred to an entity of the
+// collection being watched since the most recent call to Next
+// or the Watch call that created the srvStringsWatcher.
+func (w *srvStringsWatcher) Next() (params.StringsWatchResult, error) {
+	if changes, ok := <-w.watcher.Changes(); ok {
+		return params.StringsWatchResult{
+			Changes: changes,
+		}, nil
+	}
+	var err error
+	if e, ok := w.watcher.(hasErr); ok {
+		err = e.Err()
+	}
+	if err == nil {
+		err = apiservererrors.ErrStoppedWatcher
+	}
+	return params.StringsWatchResult{}, err
+}
